Create siglayer package, which is used by  and
